requirements.txt
jupyterlite-core==0.1.2
jupyterlab==4.0.6
jupyterlite-pyodide-kernel==0.1.2
plotly==5.17.0
grist-jupyterlab-widget==0.5.0
IPython
